+++ tests/test_transaction_async.py
@@ -1,8 +1,10 @@
+import asyncio
 import logging
 
 import pytest
 
 from psycopg import AsyncConnection, ProgrammingError, Rollback
+from psycopg._compat import create_task
 
 from .test_transaction import in_transaction, insert_row, inserted
 from .test_transaction import ExpectedException
@@ -615,3 +617,51 @@ async def test_str(aconn):
 
     assert "[IDLE]" in str(tx)
     assert "(terminated)" in str(tx)
+
+    with pytest.raises(ZeroDivisionError):
+        async with aconn.transaction() as tx:
+            1 / 0
+
+    assert "(terminated)" in str(tx)
+
+
+@pytest.mark.parametrize("what", ["commit", "rollback", "error"])
+async def test_concurrency(aconn, what):
+    await aconn.set_autocommit(True)
+
+    e = [asyncio.Event() for i in range(3)]
+
+    async def worker(unlock, wait_on):
+        with pytest.raises(ProgrammingError) as ex:
+            async with aconn.transaction():
+                unlock.set()
+                await wait_on.wait()
+                await aconn.execute("select 1")
+
+                if what == "error":
+                    1 / 0
+                elif what == "rollback":
+                    raise Rollback()
+                else:
+                    assert what == "commit"
+
+        if what == "error":
+            assert "would roll back" in str(ex.value)
+            assert isinstance(ex.value.__context__, ZeroDivisionError)
+        elif what == "rollback":
+            assert "would roll back" in str(ex.value)
+            assert isinstance(ex.value.__context__, Rollback)
+        else:
+            assert "would commit" in str(ex.value)
+
+    # Start a first transaction in a task
+    t1 = create_task(worker(unlock=e[0], wait_on=e[1]))
+    await e[0].wait()
+
+    # Start a nested transaction in a task
+    t2 = create_task(worker(unlock=e[1], wait_on=e[2]))
+
+    # Terminate the first transaction before the second does
+    await asyncio.gather(t1)
+    e[2].set()
+    await asyncio.gather(t2)