+++ docs/news.rst
@@ -18,6 +18,8 @@ Psycopg 3.0.6 (unreleased)
 - Don't raise exceptions on `ServerCursor.close()` if the connection is closed
   (:ticket:`#173`).
 - Fail on `Connection.cursor()` if the connection is closed (:ticket:`#174`).
+- Raise `ProgrammingError` if out-of-order exit from transaction contexts is
+  detected (:ticket:`#176`).
 - Add `!CHECK_STANDBY` value to `~pq.ConnStatus` enum.